Update patch set 1
Patch Set 1:

`user_visible` is a common property of all objects that resides under `id_perms`.

I feel that having this ability to hide certain rows of a table based on the prop is useful.

Changing it to target only a specific form would need some guidance as I'm not familiar with the system and how to override this component only in a specific place.
Is it like registering a copy of this template with a different id and using it? Should it be placed in `web-controller` repo then instead of the `web-core`?

Patch-set: 1